refactor: 우승자 선별 및 사용자 입력 검증 로직 파일 분리
